pytorch-blockClassification/requirements.txt
absl-py==0.10.0
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.4
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cmake==3.18.2.post1
cmaker==0.0.6
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
flake8==3.8.3
future==0.18.2
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.4.2
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
pandas==1.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydocstyle==5.1.1
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.7.1
pylint==2.6.0
pyls==0.1.6
PyOpenGL==3.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
PyYAML==5.3.1
pyzed @ file:///usr/local/zed/pyzed-3.2-cp37-cp37m-linux_x86_64.whl
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.6
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
snowballstemmer==2.0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
traitlets==5.0.4
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.1.0
